Fix isValid, add the other comparison operators
isValid was missing a check to see if all the numbers in the left and right
set are valid as well.

I am creating custom operators such as (.<) and (.>) for comparing two
surreal numbers instead of defining (<) and (>). This is because if I
define (<) and (>), Haskell will use my custom functions for behind the
scenes operations, such as the ordering of maps and sets. I want to
avoid this behavior for now, so that I can make changes without risk of
breaking the program.

I have renamed the lesser than or equal to function to (<==). I have
added the greater than or equal to function (>==). I have added the
functions for strictly lesser than (.<) and strictly greater than (.>).
I have defined the "like" operator (===), which compares the values of
two surreal numbers.