Add shadowzero as a fundamental number
"Shadowzero" is the name I have given to the incorrectly formed surreal
number with zero in its left set and zero in its right set. The "shadow"
part comes from the fact that it is not greater than, like, nor less
than zero. However, it acts like zero in that it is less than one and
greater than minus one.

It seems like every number has a shadow version of itself, not just
zero. This needs further testing!