Add unit testing with HSpec
This is exciting! I am learning how to use HSpec to test the various
functions and operators. I can test both specific edge cases and a
sample of randomly generated inputs. While I am still relatively new to
this kind of framework, it should not be difficult to incorporate into
my workflow