Create a data type for infinite sets
ISet is a way to construct infinite sets. Hopefully this will allow me
to create new interesting surreal numbers. Omega, for example, contains
all of the integers in its left set and an empty right set. Eventually
this could give us ways to add and multiply by infinity!

Currently ISet is defined as a set of qualifiers. If a number meets all
of the qualifiers, it is part of the set. The whole numbers, for
example, might be defined as [isInt, > 0].

isInt currently only works for integers made with the fromInt function.
If the number being tested is the result of the add function, for
example, it will probably not give the correct answer.